# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Standard envelope for all 4xx / 5xx responses.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ExternalIdProvider": {
        "enum": [
          "GOOGLE",
          "YELP",
          "APPLE"
        ],
        "title": "ExternalIdProvider",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganizationPlaceOwnerRead": {
        "description": "One ``place_owners`` row belonging to an org, with the place\nsummary nested inline.\n\nIncludes REVOKED history alongside live rows so the org detail\npage can show \"Acme Inc used to own Halal Test Diner\" as\nmeaningful audit context. UI sorts ACTIVE-first and can choose to\nfade/badge the REVOKED entries.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "place": {
            "$ref": "#/components/schemas/OrganizationPlaceSummary"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "status",
          "created_at",
          "place"
        ],
        "title": "OrganizationPlaceOwnerRead",
        "type": "object"
      },
...
      "PlaceAdminPatch": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "PlaceAdminPatch",
        "type": "object"
      },
      "PlaceAdminRead": {
        "description": "Admin view of a place \u2014 exposes the soft-delete fields the public\n`PlaceRead` schema deliberately hides (the public `/places/{id}` 404s\non deleted rows, so it has no reason to surface them). Used by the\nadmin browse and detail endpoints so the UI can show a \"Deleted\"\nbadge and flip between Delete / Restore actions.\n\nAlso exposes canonical address fields (city, region, country_code,\npostal_code, timezone) \u2014 populated by ``/admin/places/ingest`` from the\n``canonical_source`` provider. Admin list/detail pages use these for\nsorting, filtering, and display.",
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "canonical_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExternalIdProvider"
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_deleted": {
            "title": "Is Deleted",
            "type": "boolean"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "lat",
          "lng",
          "is_deleted"
        ],
        "title": "PlaceAdminRead",
        "type": "object"
      },
...
      "PlaceDeleteRequest": {
        "additionalProperties": false,
        "description": "Optional justification attached to a soft-delete.\n\nThe admin panel's Delete dialog collects a reason and passes it here.\nStored on the PlaceEvent audit row so the event history page makes\nclear *why* something was removed, not just *when*.\n\nOptional on the API so existing scripts, Bruno requests, and tests\nthat DELETE without a body keep working. The admin UI enforces a\nnon-empty reason at the form layer.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-form justification surfaced in the place's event history. Keep it under 500 chars; the UI collects a one-liner.",
            "title": "Reason"
          }
        },
        "title": "PlaceDeleteRequest",
        "type": "object"
      },
...
      "PlaceSearchResult": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "lat",
          "lng"
        ],
        "title": "PlaceSearchResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TrustHalal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/organizations/{org_id}/places": {
      "get": {
        "description": "List places this org owns (live + historical).\n\nCloses the user \u2194 org \u2194 place triangle visually: the org detail\npage can now show \"which places does Acme run?\" without bouncing\nthrough the places admin and filtering by owner.\n\nACTIVE rows come first; REVOKED history rows follow so admins can\nsee what the org USED to own \u2014 useful when triaging \"we never\nworked with Acme, why does our catalog show them?\"\n\nSoft-deleted places are included (the ``place.is_deleted`` flag\nlets the UI fade or badge them).",
        "operationId": "list_org_places_admin_admin_organizations__org_id__places_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizationPlaceOwnerRead"
                  },
                  "title": "Response List Org Places Admin Admin Organizations  Org Id  Places Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Org Places Admin",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/places": {
      "get": {
        "operationId": "list_places_admin_admin_places_get",
        "parameters": [
          {
            "in": "query",
            "name": "deleted",
            "required": false,
            "schema": {
              "default": "false",
              "pattern": "^(true|false|all)$",
              "title": "Deleted",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "description": "ISO-2 country code (e.g. US, GB). Case-insensitive.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-2 country code (e.g. US, GB). Case-insensitive.",
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "order_by",
            "required": false,
            "schema": {
              "default": "updated_at",
              "pattern": "^(updated_at|name|city|country)$",
              "title": "Order By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "exclusiveMinimum": 0,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaceAdminRead"
                  },
                  "title": "Response List Places Admin Admin Places Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Places Admin",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/places/{place_id}": {
      "delete": {
        "description": "Soft-delete a place.\n\nAccepts an optional ``reason`` that's logged to the place's event\nhistory alongside the DELETED event. The reason is free text; the\nadmin UI strongly encourages one but nothing enforces it server-side\n(keeps scripts and Bruno collections backward-compatible).\n\nIdempotent: deleting an already-deleted place is a no-op (no new\nevent row, no error).",
        "operationId": "delete_place_admin_places__place_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PlaceDeleteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid X-User-Id header."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Caller is authenticated but not an admin."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PLACE_NOT_FOUND",
                    "message": "Place not found"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No place with this id exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Reason failed validation (too short or too long)."
          }
        },
        "summary": "Delete Place",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "get_place_admin_admin_places__place_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceAdminRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Place Admin",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "patch_place_admin_places__place_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceAdminPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceAdminRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Place",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/places/{place_id}/claims": {
      "get": {
        "operationId": "get_place_claims_admin_admin_places__place_id__claims_get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Place Claims Admin Admin Places  Place Id  Claims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Place Claims Admin",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/places": {
      "get": {
        "operationId": "search_places_places_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lng",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "radius",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "maximum": 100000,
              "title": "Radius",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "exclusiveMinimum": 0,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaceSearchResult"
                  },
                  "title": "Response Search Places Places Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Places",
        "tags": [
          "Places"
        ]
      },
      "post": {
        "operationId": "post_place_places_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Place",
        "tags": [
          "Places"
        ]
      }
    },
    "/places/{place_id}": {
      "get": {
        "operationId": "get_place_by_id_places__place_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Place By Id",
        "tags": [
          "Places"
        ]
      }
    },
    "/places/{place_id}/claims": {
      "get": {
        "operationId": "get_place_claims_places__place_id__claims_get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Place Claims Places  Place Id  Claims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Place Claims",
        "tags": [
          "Places"
        ]
      }
    },
...
  }
}